Add prompt API and shared table schema service
