+++ BazarBin.Mcp.Server/Program.cs
@@ -1,3 +1,5 @@
+using BazarBin.Services;
+
 var builder = WebApplication.CreateBuilder(args);
 
 
@@ -7,6 +9,7 @@ builder.Services
     .WithToolsFromAssembly();
 
 
+builder.Services.AddSingleton<TableSchemaService>();
 builder.Services.AddHttpClient();
 
 var app = builder.Build();
