+++ BazarBin/Models/CsvImportSchema.cs
@@ -9,7 +9,8 @@ public sealed record CsvColumnSchema(
 public sealed record CsvImportSchema(
     string TableName,
     IReadOnlyList<CsvColumnSchema> Columns,
-    string? TableComment = null);
+    string? TableComment = null,
+    bool FirstRowIsHeader = true);
 
 public sealed record CsvImportResult(
     string FullyQualifiedTableName,
