+++ BazarBin/Models/PromptRequest.cs
@@ -1,0 +1,5 @@
+namespace BazarBin.Models;
+
+public sealed record PromptRequest(string Prompt);
+
+public sealed record PromptResponse(string PromptWithSchema, string Response);
