+++ Directory.Build.props
@@ -22,4 +22,9 @@
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.7"/>
     <PackageReference Include="ModelContextProtocol" Version="0.3.0-preview.4" />
   </ItemGroup>
+
+  <ItemGroup Condition="'$(MSBuildProjectName)' == 'BazarBin'">
+    <PackageReference Include="Microsoft.Extensions.AI" Version="9.9.0" />
+    <PackageReference Include="Microsoft.Extensions.AI.OpenAI" Version="9.9.0-preview.1.25458.4" />
+  </ItemGroup>
 </Project>
